refactor to use edit-site store
