+++ packages/edit-site/src/store/reducer.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { flow } from 'lodash';
+import { flow, filter } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -174,6 +174,30 @@ export function homeTemplateId( state, action ) {
 	return state;
 }
 
+export function documentSettingsDropdown(
+	state = { hoveredEntities: [] },
+	action
+) {
+	switch ( action.type ) {
+		case 'SETTINGS_DROPDOWN_HOVERED_ENTITIES_ADD':
+			return {
+				...state,
+				hoveredEntities: [ action.clientId, ...state.hoveredEntities ],
+			};
+
+		case 'SETTINGS_DROPDOWN_HOVERED_ENTITIES_REMOVE':
+			return {
+				...state,
+				hoveredEntities: filter(
+					state.hoveredEntities,
+					( blockId ) => blockId !== action.clientId
+				),
+			};
+	}
+
+	return state;
+}
+
 export default combineReducers( {
 	preferences,
 	deviceType,
@@ -183,4 +207,5 @@ export default combineReducers( {
 	templateType,
 	page,
 	homeTemplateId,
+	documentSettingsDropdown,
 } );
